+++ src/components/molecules/test/index.ts
@@ -1,1 +1,1 @@
-export * from "./test";
+export * from "./testComp";
